==> 01-convert.ipynb <==
# Converteer dat naar parquet

import numpy as np
import pandas as pd

def read_movies(file_name):
    movies = pd.read_table(
        file_name, encoding='utf_8', engine='python', sep='::', names=['imdb_id', 'title', 'genres']
    )
    titles = (
        movies['title']
            .str.extract(r'(?P<title>.+) \((?P<year>\d{4})\)')
            .astype({'year': np.int})
    )
    genres = (
        movies['genres']
            .str.lower()
            .str.replace('-', '_')
            .str.split('|')
    )
    return (
        movies
            .drop(columns=['title', 'genres'])
            .join(titles)
            .join(genres)
    )

def read_ratings(file_name):
    return pd.read_table(
        file_name, engine='python', sep='::', names=['user_id', 'imdb_id', 'rating', 'date_time'],
        parse_dates=['date_time'], date_parser=lambda s: pd.to_datetime(s, origin='unix', unit='s')
    )

def read_users(file_name):
    return pd.read_table(file_name, engine='python', sep='::', names=['user_id', 'twitter_id'])

movies = read_movies('MovieTweetings/latest/movies.dat')
movies

ratings = read_ratings('MovieTweetings/latest/ratings.dat')
ratings

users = read_users('MovieTweetings/latest/users.dat')
users

movies.to_parquet('data/movies.parquet')
ratings.to_parquet('data/ratings.parquet')
users.to_parquet('data/users.parquet')
