==> 02-clean.ipynb <==
# Opschonen

import numpy as np
import pandas as pd

def missing(df):
    return df[df.isna().any(axis='columns')]

def movie_url(imdb_id):
    return f'https://www.imdb.com/title/tt{imdb_id:07}/'

movies = pd.read_parquet('data/movies.parquet')
print(f'n_movies = {movies.shape[0]}')
movies.head()

ratings = pd.read_parquet('data/ratings.parquet')
print(f'n_ratings = {ratings.shape[0]}')
ratings.head()

users = pd.read_parquet('data/users.parquet')
print(f'n_users = {users.shape[0]}')
users.head()

## Movies

### Ontbrekend

mis_movies = missing(movies)
mis_movies

mis_movies['genres'].unique()

mis_movies['imdb_id'].apply(movie_url)

### Foutief

duplicates = movies[movies['imdb_id'].duplicated()]['imdb_id'].values
movies[movies['imdb_id'].isin(duplicates)]

movies = movies.drop_duplicates('imdb_id')
print(f'n_movies = {movies.shape[0]}')
movies[movies['imdb_id'].isin(duplicates)]

movies['year'].describe()

set(np.concatenate(movies['genres'].dropna().values))

## Ratings

### Ontbrekend

missing(ratings)

### Foutief

ratings[ratings.duplicated(['user_id', 'imdb_id'])]

ratings['rating'].describe()

## Users

### Ontbrekend

missing(users)

### Foutief

users[users['user_id'].duplicated()]

duplicates = users[users['twitter_id'].duplicated()]['twitter_id'].values
dup_users = users[users['twitter_id'].isin(duplicates)]
dup_users.sort_values(by=['twitter_id', 'user_id'])

dup_users.groupby('twitter_id').count().describe()

users['user_id'].is_monotonic

users = users.drop_duplicates('twitter_id')
dup_groups = [sorted(group['user_id'].values) for _twitter_id, group in dup_users.groupby('twitter_id')]
dup_reps = {dup_id: unq_id for [unq_id, *dup_ids] in dup_groups for dup_id in dup_ids}
ratings = (
    ratings
        .replace({'user_id': dup_reps})
        .drop_duplicates(['user_id', 'imdb_id'])
)

# Bewaren

movies.to_parquet('data/movies_clean.parquet')
ratings.to_parquet('data/ratings_clean.parquet')
users.to_parquet('data/users_clean.parquet')

# Conclusies

## Movies
Ontbrekend: genres 251 films (niet opgelost). Foutief: 2 dubbele films (opgelost), nog te verschijnen films (de dataset bestaat uit beoordeelde films, niet opgelost).

## Users
Foutief: 4832 dubbele gebruikers (opgelost).
