==> 04-normalize.ipynb <==
# Normaliseren

import numpy as np
import pandas as pd

def missing(df):
    return df[df.isna().any(axis='columns')]

def replace(df, s):
    return (
        df
            .drop(columns=s.name)
            .join(s)
            .reindex(columns=df.columns)
    )

movies = pd.read_parquet('data/movies_clean.parquet')
ratings = pd.read_parquet('data/ratings_clean.parquet')
users = pd.read_parquet('data/users_clean.parquet')

## Vijf klassen
Ter voorbereiding van classificatie, elf is excessief.

ratings = replace(ratings, ((ratings['rating'] - 5) / 5 * 2 + 3).round().astype(np.int))
ratings['rating'].describe()

## Doorlopende IDs
Ter voorbereiding van embedding.

### Movies

movies = movies.join(pd.Series(range(1, len(movies) + 1), movies.index, name='movie_id'))
movies = movies[['movie_id', 'imdb_id', 'title', 'year', 'genres']]
movies.head()

### Users

len(users)

users['user_id'].max()

users = users.join(pd.Series(range(1, len(users) + 1), users.index, name='user_id_new'))
users.head()

### Ratings

ratings.head()

ratings = (
    ratings
        .merge(movies[['imdb_id', 'movie_id']], 'left', 'imdb_id')
        .drop(columns='imdb_id')
)
ratings.head()

ratings = (
    ratings
        .merge(users[['user_id', 'user_id_new']], 'left', 'user_id')
        .drop(columns='user_id')
)
ratings.head()

ratings = ratings[['user_id_new', 'movie_id', 'rating', 'date_time']]
ratings.columns = ['user_id', 'movie_id', 'rating', 'date_time']
ratings.head()

Verwijder tijdelijke kolom.

users = users[['user_id_new', 'twitter_id']]
users.columns = ['user_id', 'twitter_id']
users.head()

# Bewaren

movies.to_parquet('data/movies_normalized.parquet')
ratings.to_parquet('data/ratings_normalized.parquet')
users.to_parquet('data/users_normalized.parquet')
