==> 05-augment.ipynb <==
# Verrijken

from datetime import datetime, date, timedelta
import numpy as np
import pandas as pd
from sklearn.preprocessing import MultiLabelBinarizer

def missing(df):
    return df[df.isna().any(axis='columns')]

pd.set_option('display.max_columns', None)

movies = pd.read_parquet('data/movies_normalized.parquet')
ratings = pd.read_parquet('data/ratings_normalized.parquet')
users = pd.read_parquet('data/users_normalized.parquet')

## Test gebruiker

user_id = users['user_id'].max() + 1
imdb_ids = [
    78748,   3890160, 2245084, 1596363, 120611, 83658,   88846,   381061,  88944,   468569, 
    91064,   91083,   1343092, 113568,  94625,  1454468, 2015381, 113243,  3460252, 371746, 
    5104604, 4116284, 1490017, 79470,   75860,  327597,  1568346, 76759,   107688,  88763,  
    1392190, 133093,  3659388, 1077258, 93773,  1677720, 1663202, 1189073, 84787,   116225, 
    100802,  993846,  94291,   119174,  83907,  151804,  70909,   74812,   49223,   118749
]
imdb_ratings = [
    3, 4, 3, 2, 3, 5, 3, 4, 5, 4,
    2, 5, 3, 4, 3, 4, 3, 4, 3, 3,
    4, 4, 4, 4, 4, 3, 4, 3, 1, 4,
    3, 3, 3, 4, 4, 1, 3, 4, 3, 1,
    3, 4, 4, 3, 3, 3, 2, 1, 3, 1
]

users = users.append(pd.Series([user_id, 0], users.columns), ignore_index=True, verify_integrity=True)
users[users['user_id'] == user_id]

ratings = ratings.append(
    pd.DataFrame(
        {
            'user_id': user_id,
            'movie_id': movies[movies['imdb_id'].isin(imdb_ids)].sort_values(['year', 'title'])['movie_id'],
            'rating': imdb_ratings,
            'date_time': pd.to_datetime(datetime.now().strftime('%Y-%m-%d %H:%M:%S'))
        }
    ),
    ignore_index=True, verify_integrity=True
)
ratings[ratings['user_id'] == user_id].merge(movies, 'left', 'movie_id')[['user_id', 'title', 'year', 'rating', 'date_time']]

## Genres

def fill_missing_genres(genres):
    if genres is None:
        return []
    return genres

genres_binarizer = MultiLabelBinarizer()
genres = pd.DataFrame(
    genres_binarizer.fit_transform(movies['genres'].apply(fill_missing_genres)),
    movies.index, genres_binarizer.classes_
)
movies = (
    movies
        .drop(columns='genres')
        .join(genres)
)
movies.head()

## Populariteit

today_one_year_ago = date.today() - timedelta(365)
popularity = pd.Series(
    ratings[ratings['date_time'].dt.date > today_one_year_ago].groupby('movie_id')['rating'].count(),
    name='popularity'
)
movies = (
    movies
        .merge(popularity, 'left', 'movie_id')
        .fillna({'popularity': 0})
        .astype({'popularity': np.int})
)
movies.head()

## Actieve gebruikers
Enkel gebruikers met vijfentwintig (ongeveer twee maal gemiddeld aantal beoordelingen per gebruiker) of meer beoordelingen.

ratings_user = ratings.groupby('user_id')['rating'].count()
ratings_active = ratings[ratings['user_id'].isin(ratings_user[ratings_user >= 25].index.values)]

print(f'n_ratings = {ratings_active.shape[0]}')
ratings_active.head()

movies_active = movies[movies['movie_id'].isin(ratings_active['movie_id'].unique())]
print(f'n_movies = {movies_active.shape[0]}')
movies_active.head()

users_active = users[users['user_id'].isin(ratings_active['user_id'].unique())]
print(f'n_users = {users_active.shape[0]}')
users_active.head()

movies_active = movies_active.join(
    pd.Series(range(1, len(movies_active) + 1), movies_active.index, name='movie_id_new')
)
users_active = users_active.join(
    pd.Series(range(1, len(users_active) + 1), users_active.index, name='user_id_new')
)
ratings_active = (
    ratings_active
        .merge(movies_active[['movie_id', 'movie_id_new']], 'left', 'movie_id')
        .merge(users_active[['user_id', 'user_id_new']], 'left', 'user_id')
)

cols = list(movies_active.columns)
movies_active = movies_active[[cols[-1]] + cols[1:-1]]
movies_active.columns = cols[:-1]
users_active = users_active[['user_id_new', 'twitter_id']]
users_active.columns = ['user_id', 'twitter_id']
ratings_active = ratings_active[['user_id_new', 'movie_id_new', 'rating', 'date_time']]
ratings_active.columns = ['user_id', 'movie_id', 'rating', 'date_time']

movies_active.head()

users_active.head()

ratings_active.head()

# Bewaren

movies.to_parquet('data/movies_augmented.parquet')
ratings.to_parquet('data/ratings_augmented.parquet')
users.to_parquet('data/users_augmented.parquet')
movies_active.to_parquet('data/movies_active.parquet')
ratings_active.to_parquet('data/ratings_active.parquet')
users_active.to_parquet('data/users_active.parquet')
