==> 90-predict.ipynb <==
# Voorspellen

import pickle
import numpy as np
import pandas as pd
from sklearn.preprocessing import StandardScaler
from tensorflow.keras.models import load_model

pd.set_option('display.max_columns', None)

def save(obj, path):
    with open(path, 'bw') as file:
        pickle.dump(obj, file)

def load(path):
    with open(path, 'br') as file:
        return pickle.load(file)

def drop_seen(user_id, X, ratings):
    seen_ids = ratings[ratings['user_id'] == user_id]['movie_id'].unique()
    return X[~X['movie_id'].isin(seen_ids)]

def join_user(X, user_id):
    return (
        pd
            .DataFrame({'user_id': user_id}, X.index)
            .join(X)
    )

def to_X(user_id, X, ratings):
    return join_user(drop_seen(user_id, X, ratings), user_id)

def to_inputs_cf(X):
    return [X['user_id'], X['movie_id']]

def to_inputs_hy(X):
    return [X['user_id'], X['movie_id'], X.drop(columns=['user_id', 'movie_id'])]

def predict(inputs, model):
    return inputs[1], model.predict(inputs)

def predict_rg(inputs, model):
    movie_ids, ratings = predict(inputs, model)
    return (
        pd
            .DataFrame({'movie_id': movie_ids.values, 'rating': ratings.reshape(-1)})
            .sort_values('rating', ascending=False)
    )

def predict_cl(inputs, model):
    movie_ids, ratings = predict(inputs, model)
    classes = ratings.argmax(axis=1)
    return (
        pd
            .DataFrame(
                {
                    'movie_id': movie_ids.values, # Don't reuse index of movie_ids so ratings can be joined
                    'rating': classes + 1,
                    'confidence': ratings[range(len(ratings)), classes]
                }
            )
            .join(pd.DataFrame(ratings))
            .sort_values(['rating', 'confidence'], ascending=False)
    )

def movie_url(imdb_id):
    return f'https://www.imdb.com/title/tt{imdb_id:07}/'

def join_desc(preds, movies):
    ps = preds.merge(movies, 'left', 'movie_id')
    ps = (
        ps
            .drop(columns='imdb_id')
            .join(ps['imdb_id'].apply(movie_url))
    )
    rating_cols = list(preds.columns)[1:]
    ps = ps[['title', 'imdb_id'] + rating_cols]
    ps.columns = ['title', 'movie_url'] + rating_cols
    return ps

def top(preds, movies, n=5):
    return join_desc(preds, movies).head(n)

def drop_active(X, ratings, z=3, z_left=None, z_right=None):
    ratings_movie = ratings.groupby('movie_id')['rating'].count()
    x, s = ratings_movie.mean(), ratings_movie.std()
    z_l = z_left or ((not z_right) and z) or None
    z_r = z_right or ((not z_left) and z) or None
    l = (z_l and x - s * z_l) or 1
    r = (z_r and x + s * z_r) or ratings_movie.max()
    active_ids = ratings_movie[np.logical_or(ratings_movie < l, ratings_movie > r)].index.values
    return X[~X['movie_id'].isin(active_ids)]

# Data set
Movies, ratings en users.

movies = pd.read_parquet('data/movies_active.parquet')
ratings = pd.read_parquet('data/ratings_active.parquet')
users = pd.read_parquet('data/users_active.parquet')

Features zonder gebruiker.

year_pop_transformer = load('data/train_test_active.pickle')[2]
transform_cols = ['year', 'popularity']
X = movies.drop(columns=['imdb_id', 'title'])
X = (
    pd
        .DataFrame(year_pop_transformer.transform(X[transform_cols]), X.index, transform_cols)
        .join(X.drop(columns=transform_cols))
        .reindex(columns=X.columns)
)

Features voor specifieke gebruikers.

user_id_test = users[users['twitter_id'] == 0]['user_id'].values[0]
user_id_1, user_id_2, user_id_3 = (
    ratings
        .groupby('user_id')['rating']
        .count()
        .sort_values(ascending=False)
        .iloc[:3]
)
X_test = to_X(user_id_test, X, ratings)
X_1 = to_X(user_id_1, X, ratings)
X_2 = to_X(user_id_2, X, ratings)
X_3 = to_X(user_id_3, X, ratings)

# Modellen

movies_cf_r     = load_model('data/movies_cf_r.h5')
movies_cf_r_1   = load_model('data/movies_cf_r_1.h5')
movies_cf_r_2   = load_model('data/movies_cf_r_2.h5')
movies_cf_r_d   = load_model('data/movies_cf_r_d.h5')
movies_cf_r_d_1 = load_model('data/movies_cf_r_d_1.h5')
movies_cf_r_d_2 = load_model('data/movies_cf_r_d_2.h5')
movies_cf_r_d_3 = load_model('data/movies_cf_r_d_3.h5')
movies_h_r      = load_model('data/movies_h_r.h5')
movies_h_r_1    = load_model('data/movies_h_r_1.h5')
movies_cf_c     = load_model('data/movies_cf_c.h5')
movies_cf_c_1   = load_model('data/movies_cf_c_1.h5')
movies_cf_c_2   = load_model('data/movies_cf_c_2.h5')
movies_cf_c_d   = load_model('data/movies_cf_c_d.h5')
movies_cf_c_d_1 = load_model('data/movies_cf_c_d_1.h5')
movies_cf_c_d_2 = load_model('data/movies_cf_c_d_2.h5')
movies_cf_c_d_3 = load_model('data/movies_cf_c_d_3.h5')
movies_h_c      = load_model('data/movies_h_c.h5')
movies_h_c_1    = load_model('data/movies_h_c_1.h5')

predict_cf_r     = lambda X: predict_rg(to_inputs_cf(X), movies_cf_r)
predict_cf_r_1   = lambda X: predict_rg(to_inputs_cf(X), movies_cf_r_1)
predict_cf_r_2   = lambda X: predict_rg(to_inputs_cf(X), movies_cf_r_2)
predict_cf_r_d   = lambda X: predict_rg(to_inputs_cf(X), movies_cf_r_d)
predict_cf_r_d_1 = lambda X: predict_rg(to_inputs_cf(X), movies_cf_r_d_1)
predict_cf_r_d_2 = lambda X: predict_rg(to_inputs_cf(X), movies_cf_r_d_2)
predict_cf_r_d_3 = lambda X: predict_rg(to_inputs_cf(X), movies_cf_r_d_3)
predict_h_r      = lambda X: predict_rg(to_inputs_hy(X), movies_h_r)
predict_h_r_1    = lambda X: predict_rg(to_inputs_hy(X), movies_h_r_1)
predict_cf_c     = lambda X: predict_cl(to_inputs_cf(X), movies_cf_c)
predict_cf_c_1   = lambda X: predict_cl(to_inputs_cf(X), movies_cf_c_1)
predict_cf_c_2   = lambda X: predict_cl(to_inputs_cf(X), movies_cf_c_2)
predict_cf_c_d   = lambda X: predict_cl(to_inputs_cf(X), movies_cf_c_d)
predict_cf_c_d_1 = lambda X: predict_cl(to_inputs_cf(X), movies_cf_c_d_1)
predict_cf_c_d_2 = lambda X: predict_cl(to_inputs_cf(X), movies_cf_c_d_2)
predict_cf_c_d_3 = lambda X: predict_cl(to_inputs_cf(X), movies_cf_c_d_3)
predict_h_c      = lambda X: predict_cl(to_inputs_hy(X), movies_h_c)
predict_h_c_1    = lambda X: predict_cl(to_inputs_hy(X), movies_h_c_1)

# Top tien

top_ten = lambda preds, n=10: top(preds, movies, n)

ratings.groupby('movie_id')['rating'].count().describe()

## movies_cf_r_d

top_ten(predict_cf_r_d(X_test))

top_ten(predict_cf_r_d(X_1))

## movies_h_r

top_ten(predict_h_r(X_test))

top_ten(predict_h_r(drop_active(X_test, ratings, z_left=-2.5, z_right=5)))

top_ten(predict_h_r(X_1))

top_ten(predict_h_r(drop_active(X_1, ratings, z_left=-2.5, z_right=5)))

## movies_cf_c_d

top_ten(predict_cf_c_d(X_test))

top_ten(predict_cf_c_d(X_1))

## movies_h_c

top_ten(predict_h_c(X_test))

top_ten(predict_h_c(drop_active(X_test, ratings, z_left=-2.5, z_right=5)))

top_ten(predict_h_c(X_1))

top_ten(predict_h_c(drop_active(X_1, ratings, z_left=-2.5, z_right=5)))
