==> 91-explore.ipynb <==
# Verkennen

import numpy as np
import pandas as pd
from sklearn.metrics.pairwise import cosine_similarity, euclidean_distances
from tensorflow.keras.models import load_model

def get_layer(name, model):
    return next(filter(lambda layer: layer.name == name, model.layers))

def filter_title(pattern, movies):
    return movies[movies['title'].str.contains(pattern, False)]

def cos_sim(movie_id, weights):
    sims = np.squeeze(cosine_similarity([weights[movie_id]], weights))[1:]
    return (
        pd
            .DataFrame({'movie_id': range(1, len(sims) + 1), 'cos_sim': sims, 'cos_sim_abs': np.abs(sims)})
            .sort_values('cos_sim_abs', ascending=False)
            .drop(columns='cos_sim_abs')
    )

def euc_sim(movie_id, weights):
    sims = np.squeeze(euclidean_distances([weights[movie_id]], weights))[1:]
    return (
        pd
            .DataFrame({'movie_id': range(1, len(sims) + 1), 'euc_sim': sims})
            .sort_values('euc_sim')
    )

def movie_url(imdb_id):
    return f'https://www.imdb.com/title/tt{imdb_id:07}/'

def join_desc(sims, movies):
    ss = sims.merge(movies, 'left', 'movie_id')
    ss = (
        ss
            .drop(columns='imdb_id')
            .join(ss['imdb_id'].apply(movie_url))
    )
    sim_cols = list(sims.columns)[1:]
    ss = ss[['title', 'imdb_id'] + sim_cols]
    ss.columns = ['title', 'movie_url'] + sim_cols
    return ss

def top(sims, movies, n=5):
    return join_desc(sims, movies).head(n)

def count_ratings(movie_id, ratings):
    return len(ratings[ratings['movie_id'] == movie_id])

def drop_active(sims, ratings, z=3, z_left=None, z_right=None):
    ratings_movie = ratings.groupby('movie_id')['rating'].count()
    x, s = ratings_movie.mean(), ratings_movie.std()
    z_l = z_left or ((not z_right) and z) or None
    z_r = z_right or ((not z_left) and z) or None
    l = (z_l and x - s * z_l) or 1
    r = (z_r and x + s * z_r) or ratings_movie.max()
    active_ids = ratings_movie[np.logical_or(ratings_movie < l, ratings_movie > r)].index.values
    return sims[~sims['movie_id'].isin(active_ids)]

movies = pd.read_parquet('data/movies_active.parquet')
ratings = pd.read_parquet('data/ratings_active.parquet')
users = pd.read_parquet('data/users_active.parquet')

movies_cf_r     = load_model('data/movies_cf_r.h5')
movies_cf_r_1   = load_model('data/movies_cf_r_1.h5')
movies_cf_r_2   = load_model('data/movies_cf_r_2.h5')
movies_cf_r_d   = load_model('data/movies_cf_r_d.h5')
movies_cf_r_d_1 = load_model('data/movies_cf_r_d_1.h5')
movies_cf_r_d_2 = load_model('data/movies_cf_r_d_2.h5')
movies_cf_r_d_3 = load_model('data/movies_cf_r_d_3.h5')
movies_h_r      = load_model('data/movies_h_r.h5')
movies_h_r_1    = load_model('data/movies_h_r_1.h5')
movies_cf_c     = load_model('data/movies_cf_c.h5')
movies_cf_c_1   = load_model('data/movies_cf_c_1.h5')
movies_cf_c_2   = load_model('data/movies_cf_c_2.h5')
movies_cf_c_d   = load_model('data/movies_cf_c_d.h5')
movies_cf_c_d_1 = load_model('data/movies_cf_c_d_1.h5')
movies_cf_c_d_2 = load_model('data/movies_cf_c_d_2.h5')
movies_cf_c_d_3 = load_model('data/movies_cf_c_d_3.h5')
movies_h_c      = load_model('data/movies_h_c.h5')
movies_h_c_1    = load_model('data/movies_h_c_1.h5')

# Gelijkaardigheid

top_ten = lambda sims, n=10: top(sims, movies, n)

movie_weights = get_layer('movie_emb', movies_h_r).get_weights()[0]

ratings_movie = ratings.groupby('movie_id')['rating'].count().sort_values(ascending=False)
(
    pd
        .DataFrame({'movie_id': ratings_movie.index, 'n_ratings': ratings_movie.values})
        .merge(movies[['movie_id', 'title']], 'left', 'movie_id')
        .head()
)

ratings_movie.describe()

## The Wolf of Wall Street

### Cosine

w_sims = cos_sim(15299, movie_weights)
top_ten(w_sims)

top_ten(drop_active(w_sims, ratings, z_left=-1))

### Euclid

w_sims = euc_sim(15299, movie_weights)
top_ten(w_sims)

top_ten(drop_active(w_sims, ratings, z_left=-1))

## Mad Max

filter_title(r'mad max', movies)[['movie_id', 'title', 'year']]

mm_ids = [4789, 5141, 5830, 17210]

### Cosine

mm_sims = cos_sim(mm_ids[0], movie_weights)
top_ten(mm_sims)

count_ratings(mm_sims['movie_id'].iloc[1], ratings)

count_ratings(mm_sims['movie_id'].iloc[2], ratings)

count_ratings(mm_sims['movie_id'].iloc[3], ratings)

join_desc(mm_sims[mm_sims['movie_id'].isin(mm_ids)], movies)

top_ten(drop_active(mm_sims, ratings, z_left=-1))

### Euclid

mm_sims = euc_sim(mm_ids[0], movie_weights)
top_ten(mm_sims)

join_desc(mm_sims[mm_sims['movie_id'].isin(mm_ids)], movies)

top_ten(drop_active(mm_sims, ratings, z_left=-1))

## Star Wars

filter_title(r'star wars', movies)[['movie_id', 'title', 'year']]

sw_ids = [4559, 4918, 5503, 9208, 9214, 9215, 22726, 22846, 22847]

### Cosine

sw_sims = cos_sim(sw_ids[0], movie_weights)
top_ten(sw_sims)

join_desc(sw_sims[sw_sims['movie_id'].isin(sw_ids)], movies)

top_ten(drop_active(sw_sims, ratings, z_left=-1))

### Euclid

sw_sims = euc_sim(sw_ids[0], movie_weights)
top_ten(sw_sims)

join_desc(sw_sims[sw_sims['movie_id'].isin(sw_ids)], movies)

top_ten(drop_active(sw_sims, ratings, z_left=-1))

## Alien

filter_title(r'alien', movies)[['movie_id', 'title', 'year']]

filter_title(r'prometheus', movies)[['movie_id', 'title', 'year']]

a_ids = [4713, 5949, 7395, 8743, 17493, 21855]

### Cosine

a_sims = cos_sim(a_ids[0], movie_weights)
top_ten(a_sims)

join_desc(a_sims[a_sims['movie_id'].isin(a_ids)], movies)

top_ten(drop_active(a_sims, ratings, z_left=-1))

### Euclid

a_sims = euc_sim(a_ids[0], movie_weights)
top_ten(a_sims)

join_desc(a_sims[a_sims['movie_id'].isin(a_ids)], movies)

top_ten(drop_active(a_sims, ratings, z_left=-1))
